Work/core 3774

See merge request securetrading-gl/retail-platform/retail-platform/cloud-office!385

+++ grails-app/domain/uk/co/wonderlane/wlpos/reporting/Sale.groovy
@@ -26,7 +26,7 @@ class Sale {
     String usersName
     DateTime dateCreated
 
-    int refundQuantity
+    BigDecimal refundQuantity
     BigDecimal avgCostPrice // Calculated value, marked transient below.
     BigDecimal avgRetailPrice // Calculated value, marked transient below.
     BigDecimal avgMargin // Calculated value, marked transient below.
